Итоговый вариант
